[GITHUB-3] Add thin pool capability to the role
This code adds the option to enable a "thin pool" configuration in docker.
It uses a secondary EBS volume to store the images and containers for the docker instance.
The thin pool driver does not work easily with trusty.
Asserts have been put in place to ensure the new code only runs on xenial.